+++ R/param-estimates.R
@@ -24,6 +24,7 @@
 #' * random_effect_sdse -- Standard errors to the OMEGA and SIGMA elements in standard deviation/correlation format
 #' * fixed -- TRUE if parameter is fixed, FALSE otherwise
 #' * diag -- TRUE if parameter is a diagonal element in random effect matrix, FALSE if off-diagonal, NA if parameter is a THETA
+#' * shrinkage -- Shrinkage using SD parameterization (for NONMEM this is ETAshrinkSD and EPSshrinkSD)
 #'
 #'
 #' @seealso [param_labels()] [apply_indices()]
@@ -46,7 +47,7 @@ param_estimates.bbi_nonmem_summary <- function(.summary) {
   # if Bayesian method (includes NUTS) do not return df because it is incorrect and misleading
   est_method <- .summary[[SUMMARY_DETAILS]][[SUMMARY_EST_METHOD]][[num_methods]]
   if (str_detect(est_method, "Bayesian")) {
-    stop(glue("param_estimates() is not currently implemented for Bayesian methods. Passed has final estimation method: {est_method}"), call. = FALSE)
+    stop(glue("param_estimates() is not currently implemented for Bayesian methods. `.summary` has final estimation method: {est_method}"), call. = FALSE)
   }
 
   summary_vars <- with(
@@ -82,7 +83,9 @@ param_estimates.bbi_nonmem_summary <- function(.summary) {
     tibble::as_tibble()
 
   param_df[["fixed"]] <- param_df[["fixed"]] == 1 # convert from 1/0 to T/F
-  param_df[["diag"]] <- map_lgl(param_df[[SUMMARY_PARAM_NAMES]], is_diag)
+  param_df[[SUMMARY_PARAM_DIAG]] <- map_lgl(param_df[[SUMMARY_PARAM_NAMES]], is_diag)
+
+  param_df <- add_param_shrinkage(param_df, .summary)
 
   return(param_df)
 }
@@ -90,7 +93,7 @@ param_estimates.bbi_nonmem_summary <- function(.summary) {
 
 #' Check if diagonal index or not
 #'
-#' Private helper to unpack an matrix index string like '(3,3)' is for a diagonal (i.e. if the numbers are the same)
+#' Private helper to unpack a matrix index string like '(3,3)' is for a diagonal (i.e. if the numbers are the same)
 #' @param .name A character scalar containing an index string
 #' @importFrom stringr str_replace_all str_split
 #' @keywords internal
@@ -104,4 +107,74 @@ is_diag <- function(.name) {
   }
 
   return(.ind[1] == .ind[2])
+}
+
+
+#' Add shrinkage details to param table
+#'
+#' Private helper to unpack shrinkage details and create a column on the input tibble
+#' with the shrinkage assigned to the relevant parameters.
+#' @param .param_df The parameter estimates table containing, at minimum `parameter_names` and `diag` columns.
+#' @param .summary A `bbi_nonmem_summary` object.
+#' @importFrom dplyr select filter left_join bind_rows
+#' @importFrom stringr str_detect
+#' @keywords internal
+add_param_shrinkage <- function(.param_df, .summary) {
+
+  # extract shrinkage for final estimation method
+  shk <- .summary[[SUMMARY_SHRINKAGE]]
+  shk <- shk[[length(shk)]]
+
+  if (is.null(shk)) {
+    .param_df[[SUMMARY_PARAM_SHRINKAGE]] <- NA_real_
+    return(.param_df)
+  }
+
+  # check for mixture model with multiple subpops
+  if (length(shk) != 1) {
+    warning(paste(
+      glue("When using `param_estimates()` with a mixture model (multiple subpops) the `{SUMMARY_PARAM_SHRINKAGE}` column will be all `NA`."),
+      glue("Users can manually extract shrinkage for each subpop from the `{SUMMARY_SHRINKAGE}` element of the `bbi_nonmem_summary` object."),
+      sep = "\n"
+      ))
+    .param_df[[SUMMARY_PARAM_SHRINKAGE]] <- NA_real_
+    return(.param_df)
+  }
+
+  # select the first (and only) subpop
+  shk <- shk[[1]]
+
+  # filter to only omega and sigma diagonal elements
+  diag_df <- .param_df %>%
+                filter(.data[[SUMMARY_PARAM_DIAG]]) %>%
+                select({{ SUMMARY_PARAM_NAMES }}, {{ SUMMARY_PARAM_DIAG }})
+
+  # parse shrinkage for OMEGA diagonals
+  omega_df <- filter(diag_df, str_detect(.data[[SUMMARY_PARAM_NAMES]], "OMEGA"))
+  omega_shk <- shk[[SUMMARY_SHRINKAGE_OMEGA]]
+  if (nrow(omega_df) != length(omega_shk)) {
+    stop(paste(
+      glue("Found {nrow(omega_df)} OMEGA diagonals in parameter table and {length(omega_shk)} elements in `.summary[['{SUMMARY_SHRINKAGE}']][['{SUMMARY_SHRINKAGE_OMEGA}']]`."),
+      "Summary object may be malformed."
+      ))
+  }
+  omega_df[[SUMMARY_PARAM_SHRINKAGE]] <- omega_shk
+
+  # parse shrinkage for SIGMA diagonals
+  sigma_df <- filter(diag_df, str_detect(.data[[SUMMARY_PARAM_NAMES]], "SIGMA"))
+  sigma_shk <- shk[[SUMMARY_SHRINKAGE_SIGMA]]
+  if (nrow(sigma_df) != length(sigma_shk)) {
+    stop(paste(
+      glue("Found {nrow(sigma_df)} SIGMA diagonals in parameter table and {length(sigma_shk)} elements in `.summary[['{SUMMARY_SHRINKAGE}']][['{SUMMARY_SHRINKAGE_SIGMA}']]`."),
+      "Summary object may be malformed."
+    ))
+  }
+  sigma_df[[SUMMARY_PARAM_SHRINKAGE]] <- sigma_shk
+
+  # combine shrinkage tibbles with original .param_df
+  # note the left join because all THETA rows _should_ have NA values for shrinkage
+  shk_df <- bind_rows(omega_df, sigma_df)
+  out_df <- left_join(.param_df, shk_df, by = c(SUMMARY_PARAM_NAMES, SUMMARY_PARAM_DIAG))
+
+  return(out_df)
 }